Create QUESTION 11th
Write a C program to divide the integer array into two halves using function recursion. 
Call the user-defined “divide” function recursively, with the left half of the split array being passed as an argument for that function. 
Let the recursive function to get executed until the array size becomes one. 
Count the number of iterations to reach the base condition. 
Explain the working procedure of recursive function with stack structure.